+++ examples/cases/advanced/onboarding_with_eth_account.py
@@ -5,8 +5,8 @@ from eth_account import Account
 from eth_account.signers.local import LocalAccount
 
 from examples.utils import init_env
-from x10.perpetual.accounts import StarkPerpetualAccount
-from x10.perpetual.configuration import TESTNET_CONFIG
+from x10.configuration import TESTNET_CONFIG
+from x10.core.stark_account import StarkPerpetualAccount
 from x10.perpetual.trading_client.trading_client import PerpetualTradingClient
 from x10.perpetual.user_client.user_client import UserClient
 from x10.utils.string import is_hex_string
