+++ x10/perpetual/trading_client/testnet_module.py
@@ -2,12 +2,12 @@ from typing import List, Optional
 
 import tenacity
 
-from x10.perpetual.assets import AssetOperationModel, AssetOperationStatus
-from x10.perpetual.configuration import EndpointConfig
+from x10.configuration import EndpointConfig
+from x10.models.asset import AssetOperationModel, AssetOperationStatus
+from x10.models.base import X10BaseModel
 from x10.perpetual.trading_client.account_module import AccountModule
 from x10.perpetual.trading_client.base_module import BaseModule
 from x10.utils.http import WrappedApiResponse, send_post_request
-from x10.utils.model import X10BaseModel
 
 
 class ClaimResponseModel(X10BaseModel):
@@ -42,6 +42,7 @@ class TestnetModule(BaseModule):
             account_module = self._account_module
             claim_to_check = resp.data.id
 
+            # FIXME: Remove `tenacity`?
             @tenacity.retry(
                 stop=tenacity.stop_after_delay(10),
                 wait=tenacity.wait_fixed(1),
