+++ x10/perpetual/user_client/onboarding.py
@@ -6,8 +6,8 @@ from eth_account.signers.local import LocalAccount
 from fast_stark_crypto import generate_keypair_from_eth_signature, pedersen_hash
 from fast_stark_crypto import sign as stark_sign
 
-from x10.perpetual.accounts import AccountModel
-from x10.utils.model import X10BaseModel
+from models.account import AccountModel
+from models.base import X10BaseModel
 
 register_action = "REGISTER"
 sub_account_action = "CREATE_SUB_ACCOUNT"
